requirements.txt
aiohttp[speedups]
requests
numpy
pandas
uvloop
transformers
datasets
prometheus_client
prometheus_summary
pydantic
grpcio
grpcio-tools
grpcio-health-checking
friendli-client >= 1.4.1
